# Conflicts:
#	src/main/java/ru/coffeecoders/questbot/keyboards/viewers/QuestionInfoKeyboard.java
#	src/main/java/ru/coffeecoders/questbot/services/TeamService.java

+++ src/main/java/ru/coffeecoders/questbot/entities/Question.java
@@ -15,7 +15,7 @@ public class Question {
     @Id
     @Column(name = "question_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int questionId;
+    private long questionId;
 
     @Column(name = "question")
     private String question;
@@ -35,11 +35,11 @@ public class Question {
     @Column(name = "question_group")
     private String group;
 
-    public int getQuestionId() {
+    public long getQuestionId() {
         return questionId;
     }
 
-    public void setQuestionId(int questionId) {
+    public void setQuestionId(long questionId) {
         this.questionId = questionId;
     }
